+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+
+### Breaking changes
+
+- Changed `Pwm::automatic_output_disable` to `output_disable`, to actually disable the output.
+
 ## [0.3.0]
 
 ### Breaking changes